Cleanup the warnings in gcc-4.1.0 compilation
Cleanup the warnings generated in GCC 4.1.0 compilation of kexec-tools.

This patch is created on top of the following level of
kexec-tools:

- kexec-tools-1.101.tar.gz (from eric biederman's site or
from lse site)
- kexec-tools-1.101-kdump6.patch (consolidated patch posted
on
http://lse.sourceforge.net/kdump/patches/1.101-kdump6/kexec-tools-1.101-kdump6.patch)

Review and suggestions are welcome.

Note:
Resending the patch since its not delivered to both fastboot and
linuxppc64-dev mailing list.

Regards,
Mohan.

+++ kdump/kdump.c
@@ -54,7 +54,7 @@ static void *xmalloc(size_t size)
 	result = malloc(size);
 	if (result == NULL) {
 		fprintf(stderr, "malloc of %u bytes failed: %s\n",
-			size, strerror(errno));
+			(unsigned int)size, strerror(errno));
 		exit(7);
 	}
 	return result;
